% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{estimateBLL}
\alias{estimateBLL}
\alias{estimateBLL2}
\title{Estimate B minimizing penalized logLik}
\usage{
estimateBLL(Sigma, B, C = diag(ncol(Sigma)), eps = 0.01, alpha = 0.1,
  maxIter = 1000, trace = 0, lambda = 0, r = FALSE)

estimateBLL2(Sigma, B, C = diag(ncol(Sigma)), eps = 0.01,
  alpha = 0.1, maxIter = 1000, trace = 0, lambda = 0, r = FALSE)
}
\arguments{
\item{Sigma}{the empirical covariance matrix}

\item{B}{an initial guess for matrix B}

\item{C}{the value of the noise matrix C}

\item{eps}{stopping criteria}

\item{alpha}{step size}

\item{maxIter}{maximum number of iterations}

\item{trace}{if >0 print info}

\item{lambda}{sparsity}

\item{r}{logical}
}
\value{
the estimated B matrix
}
\description{
Estimate B minimizing penalized logLik
}
