% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{estimateBP}
\alias{estimateBP}
\title{Estimate B minimizing Frobenius distance from inverse covariance}
\usage{
estimateBP(Sigma, B, C = diag(ncol(Sigma)), eps = 0.01, alpha = 0.01,
  beta = 2, maxIter = Inf, trace = 0, lambda = 0.1, r = TRUE)
}
\arguments{
\item{Sigma}{the empirical covariance matrix}

\item{B}{an initial guess for matrix B}

\item{C}{the value of the noise matrix C}

\item{eps}{stopping criteria}

\item{alpha}{step length}

\item{beta}{increasing factor for the step size}

\item{lambda}{}
}
\value{
the estimated B matrix
}
\description{
Estimate B minimizing Frobenius distance from inverse covariance
}
