% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{estimateBS}
\alias{estimateBS}
\title{Estimate B minimizing Frobenius norm}
\usage{
estimateBS(Sigma, B, C = diag(ncol(Sigma)), eps = 0.01, alpha = 0.1,
  maxIter = 1000, trace = 0, lambda = 0.1, r = TRUE)
}
\arguments{
\item{Sigma}{the empirical covariance matrix}

\item{B}{an initial guess for matrix B}

\item{C}{the value of the noise matrix C}

\item{eps}{stopping criteria}

\item{alpha}{step length}

\item{lambda}{}
}
\value{
the estimated B matrix
}
\description{
Estimate B minimizing Frobenius norm
}
