% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.R
\name{lrtB}
\alias{lrtB}
\title{Estimate triangular OU from cross sectional data}
\usage{
lrtB(data, D = diag(ncol(data)), alpha = 0.1, B = NULL, ...)
}
\arguments{
\item{data}{the sample from the invariant distribution}

\item{D}{noise matrix}

\item{alpha}{significance level for the LRT}

\item{B}{initial estimate for the coefficient matrix}

\item{...}{additional parameters passed to \code{optim}}
}
\description{
Use LRT
}
