% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{optimizeBtriang}
\alias{optimizeBtriang}
\title{Maximum likelihood optimization for triangular B}
\usage{
optimizeBtriang(B, S, C = diag(nrow(B)), ...)
}
\arguments{
\item{B}{initial coefficient matrix}

\item{S}{empirical covariance matrix}

\item{C}{positive definite term of the Lyapunov equation}

\item{...}{additional parameters passed to \code{optim}}
}
\value{
The MLE estimation of the coefficient matrix
}
\description{
Maximum likelihood optimization for triangular B
}
