% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{estimateBLL}
\alias{estimateBLL}
\alias{estimateCLL}
\alias{estimateBF}
\title{Estimate Ornstein-Uhlenbeck process}
\usage{
estimateBLL(Sigma, B, C = diag(ncol(Sigma)), eps = 0.01, alpha = 0.2,
  beta = 0.5, maxIter = 1000, trace = 0, lambda = 0, r = FALSE,
  h = FALSE, pert = FALSE)

estimateCLL(Sigma, B, C = diag(ncol(Sigma)), C0 = diag(ncol(Sigma)),
  eps = 0.01, alpha = 0.2, beta = 0.5, maxIter = 1000, trace = 0,
  lambda = 0, r = FALSE, t0 = 1)

estimateBF(Sigma, B, C = diag(ncol(Sigma)), C0 = diag(ncol(Sigma)),
  eps = 0.01, alpha = 1, maxIter = 1000, trace = 0, lambda = 0.1,
  beta = 0.5, r = FALSE, h = FALSE)
}
\arguments{
\item{Sigma}{the empirical covariance matrix}

\item{B}{an initial guess for the coefficient matrix B}

\item{C}{the initial guess for the noise matrix C}

\item{eps}{stopping criteria}

\item{alpha}{parameter backtracking}

\item{beta}{parameter backtracking}

\item{maxIter}{maximum number of iterations}

\item{trace}{if >0 print info}

\item{lambda}{penalization coefficient}

\item{r}{logical the set of parameter will be updated at every iteration}

\item{h}{logical if TRUE only the non-zero entries of B will be updated}

\item{C0}{penalization matrix}
}
\value{
the estimated B matrix (\code{estimateBLL}) or
the estiamted C matrix (\code{estiamteCLL}).
}
\description{
Estimate the coefficients and the noise term of an Ornstein-Uhlenbeck
process (\eqn{dX(t) = BX(t) + \sqrt{C} dW(t)}) using penalized maximum-likelihood.
}
